PP, quotes in messages and code refactor
Now the good pp are display per message.

If you write ' or " in your message, the message display by web socket will not display them.
This is for not break the json parser.

The code for chat are now separate into 3 files : chat.js, messages.js and communication.js

Bug discover : the pp in the profile page is gone. Come back my friend :(