GWS version 1 complete
Now the GWS (Game WebSocket) create a game, do one step and sleep 3 seconds to emulate the match.
After the math, the GWS send to all player that the game if finish. It cause to create a button to
go bakc to main page.
The GWS tell to the WS (main WebSocket) that the port is no longer in use.
	* think about how to move this atfer the real close of the socket *
After this, the GWS close it self, and the port can be user to a new game.